requirements.txt
# --- Core Data & Config ---
pandas>=2.0.0
openpyxl>=3.1.0
PyYAML>=6.0.0

# --- Machine Learning & Math ---
scikit-learn>=1.3.0
scipy>=1.10.0

# --- NLP & Deep Learning ---
spacy>=3.7.0
torch>=2.0.0
sentence-transformers>=2.5.0

# --- Visualization & Notebooks ---
matplotlib>=3.8.0
seaborn>=0.13.0
ipykernel>=6.29.0